More arch fixes
