+++ app/data_models/emotional_report.py
@@ -1,0 +1,10 @@
+from typing import Tuple
+from pydantic import BaseModel
+from enum import StrEnum
+
+class EmotionalReport(BaseModel):
+    """The model of reports of emotions."""
+    emotion: EmotionalState
+    dominant_emotion: str
+    region: RegionOfEvaluation
+    face_confidence: float
