+++ app/data_models/emotions.py
@@ -1,0 +1,13 @@
+from typing import Tuple
+from pydantic import BaseModel
+from enum import StrEnum
+
+class Emotions(StrEnum):
+    """Enumeration to outline emotion types."""
+    ANGRY = 'angry'
+    DISGUST = 'disgust'
+    FEAR = 'fear'
+    HAPPY = 'happy'
+    SAD = 'sad'
+    SURPRISE = 'surprise'
+    NEUTRAL = 'neutral'
