+++ app/data_models/facial_measurements.py
@@ -1,0 +1,10 @@
+from typing import Tuple
+from pydantic import BaseModel
+from enum import StrEnum
+
+class FacialMeasurements(BaseModel):
+    """Data class to outline coordinates and measurements of a feature."""
+    x: int
+    y: int
+    width: int
+    height: int
