+++ app/utils/utility_functions.py
@@ -21,14 +21,14 @@ from pylatex import (
 )
 
 
-LOOKED_AWAY_THRESHOLD: Final[float] = 7.5
-HAPPY_TO_ANGRY_DIFFERENCE_THRESHOLD: Final[float] = 10.0
-HAPPY_TO_SAD_DIFFERENCE_THRESHOLD: Final[float] = 10.0
-NEUTRAL_TO_SURPRISED_THRESHOLD: Final[float] = 10.0
-FEAR_THRESHOLD: Final[float] = 5.0
-HAPPY_THRESHOLD: Final[float] = 20.0
-DISGUST_ANGRY_THRESHOLD: Final[float] = 12.5
-SAD_THRESHOLD: Final[float] = 15.0
+LOOKED_AWAY_THRESHOLD: Final[float] = os.environ["LOOKED_AWAY_THRESHOLD"]
+HAPPY_TO_ANGRY_DIFFERENCE_THRESHOLD: Final[float] = os.environ["HAPPY_TO_ANGRY_DIFFERENCE_THRESHOLD"]
+HAPPY_TO_SAD_DIFFERENCE_THRESHOLD: Final[float] = os.environ["HAPPY_TO_SAD_DIFFERENCE_THRESHOLD"]
+NEUTRAL_TO_SURPRISED_THRESHOLD: Final[float] = os.environ["NEUTRAL_TO_SURPRISED_THRESHOLD"]
+FEAR_THRESHOLD: Final[float] = os.environ["FEAR_THRESHOLD"]
+HAPPY_THRESHOLD: Final[float] = os.environ["HAPPY_THRESHOLD"]
+DISGUST_ANGRY_THRESHOLD: Final[float] = os.environ["DISGUST_ANGRY_THRESHOLD"]
+SAD_THRESHOLD: Final[float] = os.environ["SAD_THRESHOLD"]
 
 
 EMOTIONS_GRAPH_INTERPRETATION: Final[dict[Emotions, float]] = {
